requirements.txt
torch>=2.0.0
torchvision>=0.15.0
clip-by-openai>=0.1.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-image>=0.19.0
tqdm>=4.60.0
streamlit>=1.15.0
